+++ BlazorSplitGrid/Extensions/EnumerableExtensions.cs
@@ -1,15 +1,9 @@
 using System.Collections;
-using BlazorSplitGrid.Models;
 
 namespace BlazorSplitGrid.Extensions;
 
-internal static class DictionaryExtensions
+internal static class EnumerableExtensions
 {
-    internal static int NextTrack(this Dictionary<string, GutterItem> self)
-    {
-        return self.Count * 2 + 1;
-    }
-
     internal static Dictionary<string, object> AddIfNotNull<T>(this Dictionary<string, object> self, string key, T? value) where T : IDictionary
     {
         if (value is null || value.Count == 0)