+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.4.0
+- Rebuilt the internal grid so it tracks more state blazor-side
+- Added ability to set/get sizes
+- Added ability to enable/disable gutters
+- Added ability to reset grid sizes
+- Updated how min/max sizes are controlled
+
 ## v1.3.0
 - Extract out base component
 - Cleared up example of using multiple rows and columns